requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Werkzeug==3.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
pdfkit==1.0.0 